First neomatrix_config merge.
Merged updated Aurora code.
config renamed neomatrix_config to avoid name conflicts
matrix_clear is now matrix->clear
Moved project specific questions from neomatrix_config.h to nfldefines.h

+++ Aurora/Effects.h
@@ -25,10 +25,10 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+#include "matrix.h"
+
 #ifndef Effects_H
 #define Effects_H
-
-#include "matrix.h"
 
 #define MAX_COLOR_VALUE 255
 
@@ -57,7 +57,7 @@ void drawForegroundHLine(int16_t x0, int16_t x1, int16_t y) {
   int i;
 
   for (i = x0; i <= x1; i++) {
-    matrix->drawPixel(i, y, 1);
+    matrix->drawPixel(i, y, (uint32_t)1);
   }
 }
